Major patch
So it seems Bamboodl didn't initialize the 'dead' URLs list in the JSON
data which causes it to crash. It *was* created by the Bamboodl init
code, but some edits in the last month changed that functionality.

Thanks to 285995 from 8/furry/ for reporting this. . .

This patch should fix things for anyone who's already started trying to
save threads to their subscribed list.